refactor: extract OrgRedirect component and fix GuardOrganization redirect
- Fix GuardOrganization hardcoding redirect to /app even when used in
  /manager routes by adding a fallbackRedirect prop
- Extract duplicated org redirect logic from app and manager index
  routes into a shared OrgRedirect component
